+++ py-packages/engine/corelle/tests/test_database_rotations.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from corelle.math.util import unit_vector
 import numpy as N
 import quaternion as Q
-from corelle.server.database import db
+from corelle.engine.database import db
 from geoalchemy2 import Geometry
 from geoalchemy2.shape import to_shape
 from geoalchemy2.elements import WKBElement
